% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{no_match_append}
\alias{no_match_append}
\title{Appends "_x" to "string" if x is not a matching string
warns about the changes}
\usage{
no_match_append(string, x)
}
\arguments{
\item{string}{a named string: string = c(opt = "value")}

\item{x}{a string}
}
\value{
"string_x"
}
\description{
Appends "_x" to "string" if x is not a matching string
warns about the changes
}
